personal median weekly income 2011 - 2021 populating; issue for 2006 TypeError
